kvm/book3s: Make kernel emulated H_PUT_TCE available for "PR" KVM
There is nothing in the code for emulating TCE tables in the kernel
that prevents it from working on "PR" KVM... other than ifdef's and
location of the code.

This renames book3s_64_vio_hv.c to book3s_64_vio.c and moves the
bulk of the code there.

This speeds things up a bit on my G5.

Signed-off-by: Benjamin Herrenschmidt <benh@kernel.crashing.org>
[agraf: fix checkpatch complaints]
Signed-off-by: Alexander Graf <agraf@suse.de>