KVM: PPC: Allow emulator to check for NV reg reads
Most of the time there's no need to store all NV registers just to handle
an intercept. For those few times where it's actually necessary, we can
live with a small slowdown, as long as the normal path is fast.

Teach the emulator about a new flag, "has_nv_regs", which it can use to
check if an emulated instruction needs non-volatile registers to work.
If so, it can bail out and thus notify the asm code that it needs to
store the NV register contents in memory.

Signed-off-by: Alexander Graf <agraf@suse.de>