KVM: PPC: Book3S: Save/Restore NV GPRs only when needed
Today, we're always saving/restoring even non-volatile registers every
time the VM exits. This is quite excessive and work that could easily
be reduce if we handle things a bit more cleverly.

So leverage the new information we finally have about register usage of
instructions to optimize the entry/exit code path, so it doesn't always
have to save/restore non-volatile registers.

Signed-off-by: Alexander Graf <agraf@suse.de>