MMC: Exynos: Add module alias for dw mmc
We are missing a module link in the dw mmc exynos driver
back to its module. This confuses the openSUSE mkinitrd
logic, so that we can not find the module as the driver for
the root device on a running system.

Add a THIS_MODULE link to the device registration, making
everyone happy again.

Signed-off-by: Alexander Graf <agraf@suse.de>