KVM: PPC: Emulate critical sections when we hit them
Usually the idea behind critical sections is that we don't ever trap in
them. However, we may not be that lucky always.

When we do hit a critical section while an interrupt is pending, we need
to make sure we can inject the interrupt right when the critical section
is over.

To achieve this, we just emulate every single instruction inside the
critical section until we're out of it.

Note: For new we don't trigger this code path until we properly emulate
      all the instructions necessary to run Linux guests well again.

Signed-off-by: Alexander Graf <agraf@suse.de>